refactor: move jwt stratege to auth root
