+++ packages/apps/reputation-oracle/server/src/modules/auth/jwt-http-strategy.ts
@@ -6,11 +6,11 @@ import {
   JWT_STRATEGY_NAME,
   LOGOUT_PATH,
   RESEND_EMAIL_VERIFICATION_PATH,
-} from '../../../common/constants';
-import { UserEntity, UserStatus, UserRepository } from '../../user';
-import { AuthConfigService } from '../../../config/auth-config.service';
-import { TokenRepository } from '../token.repository';
-import { TokenType } from '../token.entity';
+} from '../../common/constants';
+import { UserEntity, UserStatus, UserRepository } from '../user';
+import { AuthConfigService } from '../../config/auth-config.service';
+import { TokenRepository } from './token.repository';
+import { TokenType } from './token.entity';
 
 @Injectable()
 export class JwtHttpStrategy extends PassportStrategy(
