--- packages/apps/reputation-oracle/server/src/modules/auth/strategy/index.ts
@@ -1,1 +1,0 @@
-export * from './jwt.http';
